cleanup

+++ src/main/Die.java
@@ -18,12 +18,12 @@ public class Die
   {
     int output = 0;
     Random r = new Random();
-    for(int i = 1; i <= (int) dieCount; i++)
+    for(int i = 1; i <= dieCount; i++)
     {
       // Weil nextInt(int) von 0 bis nSides-1 geht.
-      output += r.nextInt((int)nSides) + 1;
+      output += r.nextInt(nSides) + 1;
     }
-    return (int) output;
+    return output;
   }
   
   /**
@@ -38,7 +38,7 @@ public class Die
   {
     int output = 0;
     Random r = new Random();
-    output += r.nextInt((int)nSides) + 1;
+    output += r.nextInt(nSides) + 1;
     dieCount -= 1;
     output += (dieCount > 0) ? rollDie_recursively(nSides, dieCount) : 0;
     return output;
